doozer/requirements.txt
aiofiles
bashlex
click >= 8.1.3
dockerfile-parse >= 0.0.13
defusedxml
future
koji
kubernetes ~= 28.0
kubernetes-asyncio ~= 28.0
openshift ~= 0.13.2
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
pip_system_certs
PyGitHub == 1.59.1
pyyaml >= 5.1
requests
requests_kerberos
semver
tenacity ~= 8.4, != 8.4.0  # https://github.com/jd/tenacity/issues/471
wrapt
mysql-connector-python >= 8.0.21
pydantic ~= 1.10.7
python-dateutil >= 2.8.1
openshift-client ~= 2.0.1
ruamel.yaml
setuptools>=70.0.0 # not directly required, pinned by Snyk to avoid a vulnerability
aiohttp
jira==3.4.1
ghapi
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
